页面采用AdminLTE模板，首页、登录、注册页面ui

+++ src/main/java/com/qimiaochong/contorller/TopicController.java
@@ -22,7 +22,7 @@ public class TopicController {
     }
 
     @GetMapping("/topic/{id}")
-    public Topic findOne(@PathVariable Long id){
+    public Topic findOne(@PathVariable Integer id){
         return topicService.findOne(id);
     }
 }
